feat(api): Add JsDocs + Small refactor
